requirements.txt
requests
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.13.24
botocore==1.15.41
celery==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.16
gunicorn==20.0.4
importlib-metadata==1.6.0
jmespath==0.10.0
kombu==4.6.10
Pillow==7.1.2
psycopg2==2.8.5
python-binary-memcached==0.29.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
uhashring==1.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
